+++ lib/src/feature_file.dart
@@ -1,5 +1,6 @@
 import 'package:bdd_widget_test/src/bdd_line.dart';
 import 'package:bdd_widget_test/src/feature_generator.dart';
+import 'package:bdd_widget_test/src/generator_options.dart';
 import 'package:bdd_widget_test/src/step_file.dart';
 
 class FeatureFile {
@@ -7,6 +8,7 @@ class FeatureFile {
     this.featureDir,
     this.package,
     this.existingSteps = const <String, String>{},
+    this.generatorOptions = const GeneratorOptions(),
     String input,
   }) : _lines = _prepareLines(input
             .split('\n')
@@ -14,8 +16,13 @@ class FeatureFile {
             .map((line) => BddLine(line))) {
     _stepFiles = _lines
         .where((line) => line.type == LineType.step)
-        .map(
-            (e) => StepFile.create(featureDir, package, e.value, existingSteps))
+        .map((e) => StepFile.create(
+              featureDir,
+              package,
+              e.value,
+              existingSteps,
+              generatorOptions.externalSteps ?? [],
+            ))
         .toList();
   }
 
@@ -23,9 +30,14 @@ class FeatureFile {
   final String package;
   final List<BddLine> _lines;
   final Map<String, String> existingSteps;
+  final GeneratorOptions generatorOptions;
   List<StepFile> _stepFiles;
 
-  String get dartContent => generateFeatureDart(_lines, getStepFiles());
+  String get dartContent => generateFeatureDart(
+        _lines,
+        getStepFiles(),
+        generatorOptions.testMethodName ?? defaultTestName,
+      );
 
   List<StepFile> getStepFiles() => _stepFiles;