+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## [0.7.0] - External steps and build.yaml settings
+
+* Implement external steps
+* Add possibility to change `testMethodName` in `build.yaml`
+
 ## [0.6.1] - Background/After fix
 
 * Fix paramteres in Background/After sections